+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `PodCreateOptsBuilder::portmappings`
 * Add `PodCreateOptsBuilder::security_opt`
 * Add `PodCreateOptsBuilder::service_container_id`
+* Add `PodCreateOptsBuilder::share_parent`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync